Make project ignore issues
Issues belong on GitHub, not in git. Sometimes it is useful to store
drafts of issues so that you can copy-and-paste snippets of code and
formulate an issue over time. Rather than logging in to GitHub to
document these changes, users may store their drafts in an issues/
directory that won't be tracked and that won't clutter up your git
stashing.

Be aware that hard resets may permanently erase drafts in the issues/
directory. Use this directory only for temporary storage, not for
permanent logging.