Add preliminary notebook for viewing spe files
A first attempt at creating an SPE class for handling spe files as
objects, with the intention of saving them as an HDF5 file for
standardized access.

This draft mixes class definitions and the first part of a
yet-to-be-completed module into the same notebook responsible for
viewing and testing the module.

Requires: winspec.py