Document Verdi standard operating procedures
Generate the .pdf version of the procedures from the .md version with
pandoc:

```bash
pandoc -t html -o verdi_SOP.pdf verdi_SOP.md
```

The inclusion of the intermediate html file is so that the grid table is
rendered properly. The grid table does not render natively in GitHub
from the .md version.