Start z-stage third-party improvements
Document a plan to use the third-party PI-micos stage with AFM and NSOM
functionality.

AFM and NSOM are available in ScanCtrlSpectroscopyPlus but depend on the
OEM stage being connected properly. The OEM stage died and no identical
replacement exists. The hardware to support AFM and NSOM still exists on
the WITec AlphaSNOM, but we'll need customized scripts to allow the
third-party stage to make use of the WITec analog signals.