Optimize datetime assignment to handle timezones
Rewrite the datetime assignment to use a simpler algorithm that doesn't
involve assigning datetime fields by sequence position in a string.

While the previous solution handles any combination of YYYY-MM-DD
format with generally unspecified separators, it cannot easily pick out
the case where a datetime sometimes has timezone information and
sometimes does not. It also does not allow for permutations of the
fields in case a user specifies a date as MM-DD-YYYY. Modifying the
solution to handle both YY-MM-DD formats and YYYY-MM-DD formats is also
not trivial.

This new solution allows for searching strings with a set of explicitly
defined formats. This requires slightly more work upfront to define the
allowed formats, but makes maintenance easier to digest if the script
needs to be upgraded further. Formats can be added to include YY-MM-DD
strings or strings with optional timezone information.