Make filename format expect measurement number
Winspec files automatically append a measurement number during
acquisition. The utility function that combines metadata from SPE and
WIP files by common filename association fails if the WIP and SPE file
do not have the same naming convention.

We can either agree to delete the measurement number from the SPE file,
or to append a measurement number to the end of a WIP filename so that
the WIP and SPE filenames match exactly.

This commit upholds the latter convention and introduces a mechanism in
the filename field sorting to automatically identify the last field as
measurement number, whereas previously it was identified as a datetime
string.