Add test coverage for filename utilities
Provide tests to validate working order of the utility function that
parses metadata from a filename.

This does not include tests for validating metadata from SPE, WIP, or
YAML files.

This test gives an error when attempting to access the winspec, utils, and dates.yaml files.
We have confirmed that the test will work if we fix the above issues.