Reword Witec class variables
magic -> file_type
dict -> contents

The extracted dictionary of data from the .WIP file is now accessed from
a wip object as wip.contents instead of wip.dict. Future refactoring
efforts aim to supply meaningful shortcuts to the nested hierarchy of
data with short attributes like wip.data and wip.info, the results of
which are sometimes also dictionaries. Using "dict" as a shortcut for
the entire .WIP dictionary is ambiguous in this case; "contents" more
clearly describes that it is the contents of the entire wip object.