Make Witec suport info function
Some fields in the data dictionary contain informational text relating
to the acquisition. This is under a specific nested key in
the "Data 1" key from the main data dictionary, but if multiple
acquisitions are stored in a single .WIP file, additional information
may be present in subsequent dictionaries.

The info() function allows quick access to the relevant nested keys
pertaining to the metadata based on a specific data number (Data 1 by
default).