Fix document Verdi SOP to update shut down temp
The offical instructions say to wait until the LBO crystal reads 40
degrees C before turning the back panel switch to OFF. This assumes a
working battery to maintain the remaining cooldown procedure on battery
power. The battery as of September 2023 does not hold a charge, so make
the recommended minimum temperature lower before resuming a complete
shut down.