Add preliminary SPE viewer notebook
Create a dedicated notebook for showcasing the SPE class attributes
available to Winspec file objects.

This notebook is subject to change when the SPE class is finalized in
its public facing methods. Note that this notebook is simpler in scope
than the previously developed spe_viewer, which attempted to link SPE
files to WIP files and show the collected data as an image.