Drop spe_viewer notebook
Abandon this notebook because we have outsourced much of the file
extraction methods to dedicated helper files. We drop this notebook
entirely instead of rewriting it because it is superseded by a newer
notebook `viewer_hdf5`; this also allows us to drop all commits that
reference this notebook in a future clean-up effort.