Make curve fitting notebook use lmfit
Use lmfit models instead of using scipy.optimize.curve_fit on hand-made
Lorentz and Voigt models. Fitting a spectrum with multiple peaks is now
easier to extend and manage for individual lineshapes without needing to
build custom loops.

The model in this commit is tailored specifically for one dataset
(hBN-10-CH4_loc-TL_X07_...532nm_..._obj-100x) and may not be applicable
to a general-purpose fit.