Make position dataset centered on origin
WITec coordinates use the origin parameter to describe the center point
of the region of acquisition. These coordinates are for the piezo stage,
which has a range of 100 µm. A coordinate value 0 µm cooresponds to the
center point of the range on the piezo drivers, so a 100 µm scan will
raster from -50 µm to + 50 µm.