Add makefile for reproducible data builds
Sync data from a network location into this repository so that
computations can be done on local files. This makes data transformations
must faster, at the expence of increased storage burden on the end user.

Pull data from a network location with the following command:

make data /path/to/network/data

This populates a data/ directory in this project with the contents of
the network folder specified above.

Run the following simplified command anytime you wish to synchronize the
data folder from the network.

make data